Solution for check

+++ src/main/java/com/zgrelle/sping_boot_jpa_repositiory_thymeleaf_security/dao/UserRepository.java
@@ -4,9 +4,8 @@ import com.zgrelle.sping_boot_jpa_repositiory_thymeleaf_security.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import org.springframework.security.core.userdetails.UserDetails;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
     @Query("select b from User b where b.userName = :name")
     User getUserByName(@Param("name") String name);
